fix: SPM build issues with Xcode 16 (issue #1793)
https://github.com/parse-community/Parse-SDK-iOS-OSX/issues/1793

There is an issue with the Package.swift file where it uses wildcard
globbing to refer to all the private header directories that should
be included when building.

It turns out that it was a mistake that Xcode ever supported wildcard
globbing in the first place, as described here:

https://forums.swift.org/t/recursive-headersearchpath/59769

To fix it, we are replacing the wildcard glob with a list of all
header directories.

This issue was first observed on Xcode 16.0.0 beta 1. I confirmed that
it now builds correctly on Xcode 16.0.0 beta 2.